Bug #18487724 SETTING DEFAULT_STORAGE_ENGINE CAN CRASH IN INTERN_PLUGIN_LOCK
When for some reason the evaluation of the SET default_storage_engine returned a NULL value 
the sys_var::check() code decided to just set the new storage engine to C NULL and continue.
This caused problems down the road.
Fixed by properly checking for a null value (val_str() returning a NULL) and propagating 
the error up.
Fixed spacing