Bug #18816897- IGNORE_SERVER_IDS NOT RESETTING POST RESET SLAVE ALL
RESET SLAVE ALL is supposed to clear everything
that is set by CHANGE MASTER TO.
Following RESET SLAVE ALL, the server should
forget that it was used as a slave. Because of
the buggy behaviour RESET SLAVE ALL did not
clear ignore_server_ids. This followed by
by a new replication set up still contained the
old list of server-ids.

Fixed by emptying the list contaning server-ids
to ignore on RESET SLAVE ALL.