Bug#18900309 PERFORMANCE SCHEMA MEMORY INSTRUMENTATION IGNORES SETUP TABLES
This is a performance bug, in pfs_memory_alloc_v1().

The recently added performance schema instrumentation for memory
does not honor:
- the per instrument enabled flag in table setup_instruments
- the consumers in table setup_consumers

As a result, memory allocation is always instrumented,
even when the instrumentation is supposed to be turned off.
This leads to un necessary statistics collection,
and to performance overhead.