Fix for bug#18993037
  NODE RESTARTS WHILE ANOTHER NODE IS ALREADY STARTING MAY CRASH 'MASTER'
      
This was a regression introduced by the fix for bug#16007980
That fix added the following code to Qmgr::failReportLab()

+  /**
+   * If any node is starting now (c_start.startNode != 0)
+   *   sendPrepFailReq to that too
+   */
+  if (c_start.m_startNode != 0)
+  {
+    jam();
+    cfailedNodes[cnoFailedNodes++] = c_start.m_startNode;
+    c_start.reset();
+  }
      
This was to ensure that the (local) knowledge of 'startNode'
was preparing to start, should not be lost in the case that it
re-failed before it had reconnected.

However, this also introduced a race condition where we either 
could, or could not, be notified through our normal 'channels'
about the re-fail of this node. Thus, the above code could
insert a duplicate of an already known failure in cfailedNodes[].

Later, this 'list of nodes' was converted into a BitMask
used in the PREP_FAILREQ-signal, and converted back into
a 'list of nodes' by Qmgr::execPREP_FAILREQ(). During this
BitMask conversion, the duplicated NodeId was eliminated.
However, **the 'noOfNodes' count is kept unchanged**.
Thus we end up with a materialized 'list of nodes' where
the size is of-by-one, and the last list item contains garbage.

This patch replaces the Qmgr internal usage of node lists with
NodeBitMask. This natively takes care of duplicate removal, and
simplifies the conversion to/from signal formats which use
similar bitmaps.

Furthermore, this patch also fixes an issue where 'c_start.m_startNode' 
could be reset to early. It should not be reset until 
Qmgr::execPREP_FAILREQ() has received a signal informing
about the failure of the actual 'startNode'.