Bug#18911119 assertion failed: no_data(nbytes) Check length for point GEOMETRY buffer to survive brutal-force truncates of geometry blobs; Correct length checking bugs in init_from_wkb functions.
Before this change, if a GEOMETRY blob containing a point is truncated by cast
or other function and the result is sent to GIS functions,it might cause
buffer overrun asserts, because we didn't expect such truncated and incomplete
blobs for points, we assumed the GEOMETRY buffer for point is always intact
and valid.
Now when scanning the point GEOMETRY blob buffer, we always do buffer length
check so that we can detect truncated/incomplete buffer and avoid buffer
overrun asserts. Also, length checking for WKB of geometry collection,
multipolygon and polygon is corrected.