New config parameter PatchConfig.
The parameter can be set for all kinds of nodes.  It is used in
the same way as NDB_PATCH, but if NDB_PATCH is set it has
precedence.

For processes with multiple nodes (cluster connections) the first
PatchConfig string found will be used, the other will be ignored
with a warning if conflicting.