WL#7894 - Improve dtrace support on Oracle Linux 6
Dtrace support is enabled and built if the build system(cmake)
detects the "dtrace" command on the Solaris, Mac OS X and FreeBSD.
As part of this WL, dtrace support is extented to the Oracle 
Linux 6 with UEK3 kernel. Now Dtrace support is enabled and built
if Oracle Linux 6 with UEK3 has real dtrace.