Fix for 'System Restart' part of Bug#19053226
AUTOTEST FAIL: TESTSRBANK -N MIX -L 300 -R 15 T1

This will also fix the AutoTest 'testSRBank -n SR ...'
However, the 'Node Restart' part of the 'Mix'
will not be fixed by this patch ... still working on that.

The comment for the these tests states: 

" Test that a consistent bank is restored after **graceful shutdown**"

However, the restart_cluster() was called with a default value
of the 'abort = true' argument, which implied a *non-gracefull shutdown*.

This resulted in that in cases where the first System Restart happened
before the first GCP had completed, we would restore empty Bank tables in lack
of a REDO log containg the initial fill of these tables. Thus, the
later verification of the table contents after restart failed.