Bug#19060799 GET RID OF DYNAMIC_ARRAY IN MY_TMPDIR
Our array abstractions should be removed, and substituted with modern data
structures.
  
Templatized vectors are:
 - type safe
 - easier to read/maintain
 - faster

Move mysys/mf_tmpdir.c to mysys_ssl (which is a misnomer, it's the C++ part of mysys)
Move mutex_lock.h to the include directory, so it can be used outside of server code.