Decrease some AutoTest max-time(out) values back to theirs original settings from 1-2 months back.
A lots of timeout were increased back then as there were
several testfailures due to a too low timeout value.

However, some of these tests also timeouts due to restart
failures which still has to be analyzed / fixed. So these test
will timeout independent of how high we set the timeout.

As the original timeout then seems to be sufficient, we
decrease it back in order to shorten the AutoTest run length.