BUG#14135002: RPL.RPL_ROTATE_PURGE_DEADLOCK FAILS WITH TIMEOUT ON PB2.
The test case makes use of the fine DEBUG_SYNC facility to synchronize
the execution of different threads.

The problem was that the test was using DEBUG_SYNC to signal the
waiting threads and then right away resetting it (SET
DEBUG_SYNC=RESET). This could in some cases clear the signal before
the waiting thread received it. In those cases the waiting thread
would remain waiting ad eternum, thus causing the test case to
timeout.

We fix this by waiting for the threads to receive the signal and only
then issue the reset.