Bug#18205019: RPM INSTALLATION GENERATES /USR/MY.CNF
A feature added to mysql_install_db in MySQL 5.6 was to generate my.cnf config file
from template. The reason was a wish to set a specific value of sql_mode option
for new installations. This is useful for many systems, however for some install
methods and layouts this was not needed, and for some, even not wanted.

Solution is add new option --keep-my-cnf to mysql_install_db.
When used mysql_install_db will not generate my.cnf from template.


Patch will also resolve issue mentioned in bug #68117 and #68318.