BUG#18844897:  ASSERT AT INFO.DRY_RUN
A test case configured with --skip-innodb and that checks whether the
slave server is able to restart and replication is able to either
restart or step into a degraded mode when using slave_master_info and
slave_relay_log_info tables on InnoDB, was failing with two
crashes. One during the shutdown and another during recovery on server
restart. The former, even though not entirely related to replication,
was triggering the latter. Since the former crash was sporadic, so was
the subsequent one. This was noticed in that test, because the test
stops and restarts the slave server several times.

The core of the problem was that the shutdown crash was causing the
server not to close the binary log properly. Then, on server restart,
while opening the the binary log, the server would find that it had
not been closed properly. Therefore, it would engage the log for
recovery and roll forward or roll back the pending prepared
transactions in the engine. This results in the server scanning the
binary log, collect xids - in this case none, since the tests was
running with MyISAM (remeber --skip-innodb?) - and call ha_recover.
But, inside, this function would run into an assert that checks
whether there are sufficient parties to resume the pending prepared
phases of the 2PC. Since the engine involved was MyISAM and no xid had
been collected and ha_recover was called regardless, this assert would
be triggered.

This patch fixes this issue by checking if there are enough 2PC parties 
involved before calling into ha_recover.