BUG#17487701 - VIEW WITH WITH CHECK OPTION ALLOWS INVALID REPLACE BUG#18286777 - REPLACE SUCCEEDS TO UPDATE JOIN VIEWS
BACKGROUND:
- BUG#17487701:
The row being replaced in a view might conflict with one or
more rows in the base table, some of which might not be
accessible by the view.
The buggy behaviour was that REPLACE deleted such rows (conflicting
rows inaccessible by the view). 

- BUG#18286777:
The DELETE statement is not allowed on views created using 
joins from the base tables according to The Manual.
Also, The Manual states that REPLACE is DELETE and INSERT.
Since the DELETE statement is not allowed on join views,
the REPLACE statement should not be allowed either. 

FIX:
Regarding fix for Bug#17487701, the row which we are trying
to REPLACE is checked if it is present in the View or not.
If the row which we are trying to REPLACE is found in the view,
REPLACE is allowed to proceed.
Otherwise, If the row is not found in the view and we are
trying to REPLACE it, we throw an error saying that
REPLACE cannot be executed as it requires to delete rows 
which are not in the view.
Regarding fix for Bug#18286777, before executing REPLACE,
we check if a view is a JOIN view or multitable view, we
reject the REPLACE statement.

NOTE TO DOCUMENTATION:
As noted above, the REPLACE statement might result in deletion
of more than one row before a new row is inserted. That happens
if the table contains multiple unique indexes and the new row
duplicates values for different existing rows in different
unique indexes.

The REPLACE statement will throw an error if the row being
deleted is not accessible in the view. The thing is that the
REPLACE statement might encounter inaccessible row after the
statement has already deleted a few accessible rows.
Raising an error will trigger a rollback of the current
transaction, so transactional storage engines (InnoDB, ...)
are able to rollback intermediate changes made by REPLACE.
Non-transactional storage engines (MyISAM, ...) however
are not able to rollback intermediate changes, so REPLACE
might leave the database in an inconsistent state.

Also The patch is backward incompatible: in the past REPLACE
might succeed for join view, after the patch REPLACE is always
rejected for such views.