This is a patch for bug#19022040 (SPORADICAL FAILURE IN THE TC_LOG_MMAP UNIT TEST)
The test tc_log_mmap started sporadically failed after patch for Bug#18913949
has been applied.

The reason for sporadicall failure was that we construct several instances of
TC_Log_MMap_thread that ran concurrently and used the same
instance of TCLogMMapTest. This means that they also used the same THD,
the same Transaction_ctx and the same Transaction_ctx::m_mem_root.
Therefore it was possible to have a call for free_root for the some THD's
memroot while this memroot is used in another thread. It could lead to
test failure.