WL#6837: GCS Replication: Online distributed recovery Post fix: RESET MASTER command support
Whenever a reset command was issued on a node, on restart, if data
existed in the certification relay log of this server, recovery would
hang forever.
As a fix to solve this, if the node now detects that its gtid executed set
was deleted when starting the applier module sql thread, the relay
logs are purged.
This grants the user a indirect way to reset the node completely.