Bug#16953758: PREPARED STATEMENT IS WRITTEN TO GENERAL QUERY LOG AFTER ITS EXECUTION IS FINISH
When general-logging, the "Execute" line for prepared
statements was always written after that execution.

The Execute line will be written before execution, same
as for other statements, for consistency, and so if
something crashes, we'll be able to more easily tell
what it was.

audit_plugin.test: Since general-logging (and therefore
the call to audit plugins) happens before execution now,
the SHOW counter will be one higher now (it "sees itself",
so to say). Expected behaviour, test adjusted.