BUG#16409385: SEVERAL RPL TESTS FAIL ON SOLARIS WITH TIMEOUT FOR 'CHECK WARNING'
The check-testcase stage is timing out for these two test cases
on PB2. The tests output too many spurious warnings related
to restarting the slave threads.

To fix this, we suppress warnings from the error log for 
rpl_typeconv and rpl_show_slave_status_nonblocking. These
warnings are not the focus of the test cases anyway. This reduces
the size of the error log and makes the tests less vulnerable to
fail on the check-testcase stage. Note that, error entries are 
still logged.

Also, increased the timeout for rpl_typeconv as well.