Bug #19193927 	TC TAKEOVER TAKES TOO LONG CAUSING MANY PROBLEMS
The fix for (18069334 POOR MULTI-TC-INSTANCE + MULTI NODE FAILURE HANDLING LEADS TO CLUSTER CRASH) 
introduced a new bug where the number of TC instances in the failed
node is set to an excessively high value.  This results in many rounds
of the TC failover protocol running, which takes a long time, and
impacts on the processing of other node failures, and GCP handling with
various side-effects.

This patch fixes the problem.