Bug #17780517 5.6 PARTITIONS USE MUCH MORE MEMORY THAN 5.1
PROBLEM
-------
More memory usage was seen due to increase in size of
ha_innobase object . The two character arrays 
srch_key_val*  in ha_innobase class each contribute 
approximately 6Kbytes ,so when large databases with 
many partitions are used ,memory usage increases.

FIX
---
MySQL stores the INTs datatypes in little endian format
where as innodb stores them in big endian format.The 
srch_key_val* arrays are only needed when we are 
converting INTs datatypes in the search key from little 
endian format to big endian format. So we allocate 
memory equivalent to number of INTs in each index of the 
table ,which will reduce memory footprint. Also we have
moved the allocation of the memory to ha_innobase::open()
so that the memory is only allocated during opening of 
table handler.  

[Approved by Mattiasj and Jimmy #rb5565]