BUG#18675237- NOW() : MISSING FRACTIONAL SECONDS FROM A VIEW
Analysis:
--------

Creating a view using the time function NOW(6) registers
the view query definition as NOW() in the FRM file for
the view. Thus the fractional seconds part is missing
in the view definition.

CREATE VIEW v1 AS SELECT NOW(6) as fld1;

Is transformed as:
CREATE ALGORITHM=UNDEFINED DEFINER=`root`@`localhost`
SQL SECURITY DEFINER VIEW `v1` AS select now() AS `fld1`

While registering the query definition in the FRM file,
the 'decimals'(fractional seconds) was ignored. Hence NOW(6)
was printed as NOW() in the FRM file, thus missing the
fractional seconds.

Fix:
---
While registering the view query definition in the FRM file
which uses temporal functions like NOW(), CURTIME() and
SYSDATE(), print the 'decimals'(fractional seconds part)
value, if it is set.