BUG#18472603 SQL THREAD COMMITS PARTIAL TRANSACTION WITH GTID AND              AUTO POSITIONING ENABLED
Problem:
=======

When using GTID with auto positioning, if the IO thread restarts it
might be stopped in the middle of a transaction and will ask the master
for the whole transaction again.

When the SQL thread reads this stream of events, it should rollback
the partial transaction before starting the complete one, but it is
only rolling back the previous GTID set, and setting the new GTID,
letting the partial transaction untouched.

The effect of this problem is that the SQL thread is committing the
partial transaction left on the relay log before IO thread restart
once it reaches the "BEGIN" of the transaction right after the partial
transaction, consuming the GTID.

When trying to commit the complete transaction, the SQL thread is
stopping with ER_GTID_NEXT_TYPE_UNDEFINED_GROUP, because the GTID was
already consumed by the partial transaction.

By restating the SQL thread, it will try to apply the complete
transaction again from the beginning. Unfortunately, SQL thread will
silently skip the complete transaction as the GTID is already logged.


Analysis:
========

There were two issues related to the reported problem: one related to
master's dump thread and how it clears the "created" variable of it's
FD event before sending it to the slave; other related to SQL thread
rolling back partial transactions relying only on master's FD event
"created" variable.

Master's dump thread clearing FD's "created" variable
-----------------------------------------------------

When not using GTID protocol, the dump thread sends the original
master's binlog FD to the slave with it's original "created" variable
set only if the slave asked for FD event position (4).

If the slave asked for a position greater than 4, the dump thread
assumes that it must clear the "created" variable before sending the
FD event to the slave.

So, the slave will only have this "created" set in master's FD event
if the master has just (re)started.

Assuming that if the master sent partial transaction before
"crashing/recovering" and it also lost temporary tables, the "created"
variable set in master's FD event will instruct the slave do two tasks:

a) Rollback current transaction (if there is one);
b) Drop the temporary tables;

For GTID protocol, the master was always sending the original FD of the
binlog. So, if a slave reconnected asking for transactions in the first
master binlog file, it was receiving an FD event with the "created"
variable set, making slave to rollback the current transaction (not a
problem here, as the whole transaction would appear again in the relay
log) and dropping the temporary tables (this was a problem).

SQL thread rolling back partial transactions
--------------------------------------------

SQL thread relied on master's FD "created" variable to rollback partial
transactions left on the relay log by a master crash, but there was no
mechanism to do the same for partial transactions left on the relay log
by the GTID protocol.

The "created" flag on master's FD seemed to be an effective way but
worked only when the master was in the first binlog after it's startup.

After the first binlog rotation, and before restarting the server
again, all master FD's will not have the "created" variable set. So, a
slave reconnecting and asking for transactions to any master binlog
file but the first didn't rolled back the partial transactions left
by GTID protocol.

The only effective way to SQL thread know that the IO thread has left
partial transactions in the relay log when using GTID protocol is by
verifying if the SQL thread has a partial transaction when applying a
Gtid_log_event. In this case, the SQL thread should rollback the whole
partial transaction, and not only the GTID.


Fix:
===

Master's dump thread must clear "created" variable on the FD event
before sending it to slave if the slave already has the first
transaction in the masters binlog file.

SQL thread cannot rely only in master's FD event to rollback partial
transactions left by GTID protocol on relay log. It must also rely on
Gtid_log_events being applied before committing transactions. 

@ sql/binlog.cc / sql/binlog.h

Changed read_gtids_from_binlog() and find_first_log_not_in_gtid_set()
to return the first GTID found on the binlog file if asked.

@ sql/rpl_master.cc

Added code to mysql_binlog_send() to obtain the first GTID of the
binlog file to be sent to slave, and to verify if slave already has
this first GTID before sending the FD event, cleaning the "created"
variable in this case.

@ sql/rpl_slave.cc

Added a debug point to help test case design.

@ sql/log_event.cc

Added code to make the Gtid_log_event to rollback the incomplete
transaction before rolling back the GTID.