Bug#18914499 SERVICE START FAILING WITH YUM REPO RPMS FAILING WITH CAN'T LOCATE MYSQLD ERROR
Problem: The mysqld process is started using execve which in later kernels have been patch to prevent it from changing uid/gid. Even if mysqld is started with euid=0 it will fail to downgrade its euid.

Fix: Instead this must be done by the parent process. In addition the mysqld process should never attempt to change its euid so we can't pass the --user option.