WL#7060 - Gcs plugin implementation
This is a post push fix for the shutdown problem of the plugin which
was causing test failure. The problem here is that during shutdown 
without a stop gcs_replication, the applier thread continues to be 
there and this stalls the shutdown process. 

In order to fix the problem the plugin check during the before_shutdown
hook to see if the plugin is installed and kill the applier module if the 
check returns true.