Bug#19164038 - INNODB: CLUSTER INDEX BUILD BY BULK LOAD WHEN SKIP SORT SHOULD RELEASE LATCHES
When we do table rebuild and we can skip sort for cluster index build
(rb#4788), if there is a spatial index, we need the cluster index BtrBulk
object to release latches once it finishes a batch of insert, because in
spatial index build, we may call 'log_free_check'.

rb://6087 approved by Jimmy.