WL#7688 addendum
Fix a problem when compiling with Sun Studio:
The initialization of the Log class it was passing 
a pointer to the (yet uninitialized) member m_buffer
to the base class initializer.
Fixed by passing a NULL pointer to the constructor
and in the constructor body (where m_buffer is already
initialized) calling the init() function of the base class.

Also fixed a warning of an "out" local variable 
overshadowing a global by remaming it to sout.