BUG#18710853: QUERY CACHE NOT INVALIDATED ON CASCADE DELETE               IF DB NAME HAS SPECIAL SYMBOLS
Analysis      
--------
The query cache is not invalidated for a table when the CASCADE
DELETE/UPDATE referential constraint is specified and the
database name or table name contains special characters.
      
InnoDB triggers invalidation of the query cache while performing
the check for CASCADE DELETE/UPDATE referential constraint. InnoDB
passes the key in the format of 'dbname\0tablename' to the query cache
interface where the database name and table name are in the canonical
format(encoded-format for special characters). The key used by the query
cache interface is 'dbname\0tablename' in its non-canonical format. 
The lookup performed for query cache invalidation fails for the condition
specified above due to the mismatch in the key.
      
Hence the records fetched with the query cache and without the query
cache differs.
      
Fix
---
   
Innodb now passes the key 'dbname\0tablename' in its non-canonical format to 
the query cache interface. Thus the query cache look up succeeds and the
query is invalidated.