Bug#17867478 BINLOG.BINLOG_SERVER_START_OPTIONS FAILING              DUE TO ERROR IN SERVER LOGS
Problem:

An error in the logs is causing the test to fail

Analysis:

If one has a closer look at the issue has always been there.
But since mysqld was waiting for an error, the test didn't
failed when mysqld crashed because a the missing property.

But i went further on the analysis and the test was not
testing what was coded, which was
"if the param 'datadir' is empty then we should switch
its value to point to the actual 'datadir'"

This possibly corrects also other bugs.

Fix:

Rewrite the mysql test to check if the modification of
the variable value suceeded only on the context of
rpl_make_log_name.